+++ src/components/abstract-component.js
@@ -1,4 +1,4 @@
-import {createElement} from "../utils/render";
+import {createElement} from '../utils/render';
 
 export class AbstractComponent {
   constructor() {
