+++ src/components/abstract-smart-component.js
@@ -1,4 +1,4 @@
-import {AbstractComponent} from "./abstract-component";
+import {AbstractComponent} from './abstract-component';
 
 export class AbstractSmartComponent extends AbstractComponent {
   recoveryListeners() {
