+++ src/components/no-data.js
@@ -1,4 +1,4 @@
-import {AbstractComponent} from "./abstract-component";
+import {AbstractComponent} from './abstract-component';
 
 const createNoDataComponent = () => {
   return (
