+++ src/components/sort.js
@@ -1,5 +1,5 @@
 import {AbstractComponent} from './abstract-component';
-import {SORT_ITEM_NAMES, SortType} from './../constants';
+import {SORT_ITEM_NAMES, SortType} from '../constants';
 
 const createSortItemComponent = (type, isActive) => {
   const activeClass = isActive ? `sort__button--active` : ``;
